feat: PhysX (#399)
* feat: add physx

* feat/scene/physics: add physx singleton

* chore/proj: update dependencies and physx library copy

md

proj: remove post-build script

* chore: move lib comment to cpp

* git/action: setup physx

quotation

powershell format

typo and update codeql version

missing quotation termination

physx msbuild argument

wait for cmd completion

wait does not considered as arguments

specifiy project file name

* feat/scene: expose physx scene, initialize remaining

* feat/physics: add additional physx singletons

* feat/mesh: build geometry data while loading mesh

* feat/collider: initialize static rigidbody

* feat: event delegation

* feat: object broadcasting component addition

* feat/collider: update shape if render component changes

* feat/delegate: allow to take non-member function

* feat/collider: listen render component shape changes

todo: unsubscribe if render component is removed

* fix/delegate: use function pointer as key

* fix/collider: subscribe render component if exists

* fix/collider: stock vertices branching

* fix/physics: cleanup order, pointer with nullptr

* misc: replace avx condition from stl to const value

* proj: add delegate header and code

* bug/physics: zero scalar for quaternion

* bug/constraint: wrong position used for collision resolving

penetration evaluated in each object position

* bug/constraint: collision point evaluated wrong

ray test does not present value properly (collision point floats around radius range)

* refactor/collider: replace dispatch componenet function to delegation

* fix/application: update scene initially

* feat/mesh: expose physx mesh

to create geometry structure for scaling

* feat/collider: update physx world pose

shape need to be rebuilt if scaling has changed

* bug/constraint: object accelerated without acceleration

* feat: physx simulation callback

* feat/scene: apply simulation event callback

* fix/delegate: remove force move

* fix/collider: note collider pointer as user data in physx rigid actor

* feat/todo/mesh: evaluate sdf value for collision

* refactor: split collision and simulation

* refactor/callback: move callback to global variable

* test/todo: change static body to dynamic body, exclusive shape, testing  grpi[ fo;ters

* todo/test: set all ok filter for scene, add custom event callback

* feat/collider: add stock shape for collider

* feat/collision/physx: custom simulation filter callback

records pair id and actors

* fix/collisionDet: friends simulation filter callback

* refactor/collider: wrap local function for duplicated code

* fix/scene/physx: additional flag and custom simulation filter callback

* refactor/mesh: remove physx mesh geometry

* feat/refactor/colisionDet: move collision group flag, disable internal

* fix/physxCallback: move inline global variable, use collision map direct

* bug/physx: cooking mesh with building gpu data

* fix/scene/physx: add simulation callback

* chore: move member function binder to global

* feat/rigidbody: share collider physx rigidbody

* feat/collider: reset physx if only rigidbody is removed but collider

* chore: PHYSX_ENABLED macro fix

* feat/helper: update function for imgui helpers

* feat/physics: update position from physx

* feat/proj/physx: enable unity build, physx helper

use unity build for client

* feat/chore: size check function

* refactor/physx: use mesh cooking wiht physx helper

new sdf from outside

* fix/collider: tighten imgui selection for shape

* chore: fix typecastings

collider type castings

* misc/proj: remove pch header from imgui

* bug/scene: find object does not fully checking layers

* fix/renderComp: return path type variable

* fix/collider: throw vertices getter if type is unknown

* feat: const member function binder

* feat/delegate: managing with smart pointer, copy pointer in move ctor

* fix/rigidbody: remove from owner if collider has been removed

todo/refactor: component dependency

* fix/object: missing pointer casting in RemoveComponent

* fix/collider: reset shape value if model renderer has been removed

* proj: add boost-align

* fix/type: misimplemented alignment setup

* perf/layer: use aligned vector for storage

* perf/renderer: copy instance by pointer

* fix/collider: update exclusive shape filter from collision detector

* fix/delegate: more generic and fix not working indice generating

* feat/collisionDet: update collider shape if layer mask has changed

+++ Engine/egAnimationsTexture.cpp
@@ -100,7 +100,7 @@ namespace Engine::Resources
 		const std::vector<StrongBoneAnimation>& anims, std::vector<std::vector<std::vector<Matrix>>>& preEvaluated
 	)
 	{
-		const UINT anim_count  = anims.size();
+		const UINT anim_count  = static_cast<UINT>(anims.size());
 		UINT       frame_count = 0;
 		UINT       bone_count  = 0;
